requirements.txt
# Django core
Django==4.2
djangorestframework==3.16.1

# Database
mysqlclient==2.2.7
mysql-connector-python==9.4.0

# Utilities
python-decouple==3.8
requests==2.32.5
PyJWT==2.3.0
Unidecode==1.3.6
tzdata==2025.2

# Django extensions
django-extensions==3.2.3
django-filter==23.1
django-cors-headers==4.0.0
django-multiupload==0.6.1
django-template-partials==25.1
django-colorfield==0.8.0

# Web/template
Jinja2==3.1.6
MarkupSafe==3.0.2
Markdown==3.4.3

# PDF/Image
pdfkit==1.0.0
Pillow==11.3.0

# Networking / encoding
charset-normalizer==3.4.3
idna==3.10
certifi==2025.8.3
urllib3==2.5.0
asgiref==3.9.1

# Tests
pytest==9.0.1
pytest-django==4.11.1

# Production / Deployment
gunicorn==23.0.0